requirements.txt
gradio
transformers
torch
yt_dlp
ffmpeg-python